refactor: split kanata.rs file
