+++ src/kanata/mod.rs
@@ -5,7 +5,6 @@ use log::{error, info};
 
 use crossbeam_channel::{Receiver, Sender, TryRecvError};
 use std::collections::{HashMap, HashSet};
-use std::convert::TryFrom;
 use std::io::Write;
 use std::net::TcpStream;
 use std::path::PathBuf;
@@ -23,14 +22,6 @@ use crate::{cfg, ValidatedArgs};
 
 use kanata_keyberon::key_code::*;
 use kanata_keyberon::layout::*;
-
-#[cfg(target_os = "windows")]
-#[derive(Debug, Clone, Copy, PartialEq)]
-pub enum AltGrBehaviour {
-    DoNothing,
-    CancelLctlPress,
-    AddLctlRelease,
-}
 
 pub struct Kanata {
     pub kbd_in_path: PathBuf,
@@ -50,11 +41,14 @@ use once_cell::sync::Lazy;
 static MAPPED_KEYS: Lazy<Mutex<cfg::MappedKeys>> = Lazy::new(|| Mutex::new([false; 256]));
 
 #[cfg(target_os = "windows")]
-static PRESSED_KEYS: Lazy<Mutex<HashSet<OsCode>>> = Lazy::new(|| Mutex::new(HashSet::new()));
-
+mod windows;
 #[cfg(target_os = "windows")]
-static ALTGR_BEHAVIOUR: Lazy<Mutex<AltGrBehaviour>> =
-    Lazy::new(|| Mutex::new(AltGrBehaviour::DoNothing));
+pub use windows::*;
+
+#[cfg(target_os = "linux")]
+mod linux;
+#[cfg(target_os = "linux")]
+pub use linux::*;
 
 impl Kanata {
     /// Create a new configuration from a file.
@@ -412,250 +406,12 @@ impl Kanata {
             panic!("processing loop encountered error {:?}", err)
         });
     }
+}
 
-    /// Enter an infinite loop that listens for OS key events and sends them to the processing
-    /// thread.
-    #[cfg(target_os = "linux")]
-    pub fn event_loop(kanata: Arc<Mutex<Self>>, tx: Sender<KeyEvent>) -> Result<()> {
-        info!("Kanata: entering the event loop");
-        {
-            let mut mapped_keys = MAPPED_KEYS.lock();
-            *mapped_keys = kanata.lock().mapped_keys;
-        }
-
-        let kbd_in = match KbdIn::new(&kanata.lock().kbd_in_path) {
-            Ok(kbd_in) => kbd_in,
-            Err(e) => {
-                bail!("failed to open keyboard device: {}", e)
-            }
-        };
-
-        loop {
-            let in_event = kbd_in.read()?;
-            log::trace!("{in_event:?}");
-
-            // Pass-through non-key events
-            let key_event = match KeyEvent::try_from(&in_event) {
-                Ok(ev) => ev,
-                _ => {
-                    let mut kanata = kanata.lock();
-                    kanata.kbd_out.write(in_event)?;
-                    continue;
-                }
-            };
-
-            // Check if this keycode is mapped in the configuration. If it hasn't been mapped, send
-            // it immediately.
-            let kc: usize = key_event.code.into();
-            if kc >= cfg::MAPPED_KEYS_LEN || !MAPPED_KEYS.lock()[kc] {
-                let mut kanata = kanata.lock();
-                kanata.kbd_out.write_key(key_event.code, key_event.value)?;
-                continue;
-            }
-
-            // Send key events to the processing loop
-            if let Err(e) = tx.send(key_event) {
-                bail!("failed to send on channel: {}", e)
-            }
-        }
-    }
-
-    /// Initialize the callback that is passed to the Windows low level hook to receive key events
-    /// and run the native_windows_gui event loop.
+fn set_altgr_behaviour(_cfg: &cfg::Cfg) -> Result<()> {
     #[cfg(target_os = "windows")]
-    pub fn event_loop(kanata: Arc<Mutex<Self>>, tx: Sender<KeyEvent>) -> Result<()> {
-        // Display debug and panic output when launched from a terminal.
-        unsafe {
-            use winapi::um::wincon::*;
-            if AttachConsole(ATTACH_PARENT_PROCESS) != 0 {
-                panic!("Could not attach to console");
-            }
-        };
-        native_windows_gui::init()?;
-        {
-            let mut mapped_keys = MAPPED_KEYS.lock();
-            *mapped_keys = kanata.lock().mapped_keys;
-        }
-
-        let (preprocess_tx, preprocess_rx) = crossbeam_channel::bounded(10);
-        start_event_preprocessor(preprocess_rx, tx);
-
-        // This callback should return `false` if the input event is **not** handled by the
-        // callback and `true` if the input event **is** handled by the callback. Returning false
-        // informs the callback caller that the input event should be handed back to the OS for
-        // normal processing.
-        let _kbhook = KeyboardHook::set_input_cb(move |input_event| {
-            if input_event.code as usize >= cfg::MAPPED_KEYS_LEN {
-                return false;
-            }
-            if !MAPPED_KEYS.lock()[input_event.code as usize] {
-                return false;
-            }
-
-            let mut key_event = match KeyEvent::try_from(input_event) {
-                Ok(ev) => ev,
-                _ => return false,
-            };
-
-            // Unlike Linux, Windows does not use a separate value for repeat. However, our code
-            // needs to differentiate between initial press and repeat press.
-            log::debug!("event loop: {:?}", key_event);
-            match key_event.value {
-                KeyValue::Release => {
-                    PRESSED_KEYS.lock().remove(&key_event.code);
-                }
-                KeyValue::Press => {
-                    let mut pressed_keys = PRESSED_KEYS.lock();
-                    if pressed_keys.contains(&key_event.code) {
-                        key_event.value = KeyValue::Repeat;
-                    } else {
-                        pressed_keys.insert(key_event.code);
-                    }
-                }
-                _ => {}
-            }
-
-            // Send input_events to the processing loop. Panic if channel somehow gets full or if
-            // channel disconnects. Typing input should never trigger a panic based on the channel
-            // getting full, assuming regular operation of the program and some other bug isn't the
-            // problem. I've tried to crash the program by pressing as many keys on my keyboard at
-            // the same time as I could, but was unable to.
-            try_send_panic(&preprocess_tx, key_event);
-            true
-        });
-
-        // The event loop is also required for the low-level keyboard hook to work.
-        native_windows_gui::dispatch_thread_events();
-        Ok(())
-    }
-}
-
-#[cfg(target_os = "windows")]
-fn try_send_panic(tx: &Sender<KeyEvent>, kev: KeyEvent) {
-    if let Err(e) = tx.try_send(kev) {
-        panic!("failed to send on channel: {:?}", e)
-    }
-}
-
-#[cfg(target_os = "windows")]
-fn start_event_preprocessor(preprocess_rx: Receiver<KeyEvent>, process_tx: Sender<KeyEvent>) {
-    #[derive(Debug, Clone, Copy, PartialEq)]
-    enum LctlState {
-        Pressed,
-        Released,
-        Pending,
-        PendingReleased,
-        None,
-    }
-
-    std::thread::spawn(move || {
-        let mut lctl_state = LctlState::None;
-        loop {
-            match preprocess_rx.try_recv() {
-                Ok(kev) => match (*ALTGR_BEHAVIOUR.lock(), kev) {
-                    (AltGrBehaviour::DoNothing, _) => try_send_panic(&process_tx, kev),
-                    (
-                        AltGrBehaviour::AddLctlRelease,
-                        KeyEvent {
-                            value: KeyValue::Release,
-                            code: OsCode::KEY_RIGHTALT,
-                            ..
-                        },
-                    ) => {
-                        log::debug!("altgr add: adding lctl release");
-                        try_send_panic(&process_tx, kev);
-                        try_send_panic(
-                            &process_tx,
-                            KeyEvent::new(OsCode::KEY_LEFTCTRL, KeyValue::Release),
-                        );
-                        PRESSED_KEYS.lock().remove(&OsCode::KEY_LEFTCTRL);
-                    }
-                    (
-                        AltGrBehaviour::CancelLctlPress,
-                        KeyEvent {
-                            value: KeyValue::Press,
-                            code: OsCode::KEY_LEFTCTRL,
-                            ..
-                        },
-                    ) => {
-                        log::debug!("altgr cancel: lctl state->pressed");
-                        lctl_state = LctlState::Pressed;
-                    }
-                    (
-                        AltGrBehaviour::CancelLctlPress,
-                        KeyEvent {
-                            value: KeyValue::Release,
-                            code: OsCode::KEY_LEFTCTRL,
-                            ..
-                        },
-                    ) => match lctl_state {
-                        LctlState::Pressed => {
-                            log::debug!("altgr cancel: lctl state->released");
-                            lctl_state = LctlState::Released;
-                        }
-                        LctlState::Pending => {
-                            log::debug!("altgr cancel: lctl state->pending-released");
-                            lctl_state = LctlState::PendingReleased;
-                        }
-                        LctlState::None => try_send_panic(&process_tx, kev),
-                        _ => {}
-                    },
-                    (
-                        AltGrBehaviour::CancelLctlPress,
-                        KeyEvent {
-                            value: KeyValue::Press,
-                            code: OsCode::KEY_RIGHTALT,
-                            ..
-                        },
-                    ) => {
-                        log::debug!("altgr cancel: lctl state->none");
-                        lctl_state = LctlState::None;
-                        try_send_panic(&process_tx, kev);
-                    }
-                    (_, _) => try_send_panic(&process_tx, kev),
-                },
-                Err(TryRecvError::Empty) => {
-                    if *ALTGR_BEHAVIOUR.lock() == AltGrBehaviour::CancelLctlPress {
-                        match lctl_state {
-                            LctlState::Pressed => {
-                                log::debug!("altgr cancel: lctl state->pending");
-                                lctl_state = LctlState::Pending;
-                            }
-                            LctlState::Released => {
-                                log::debug!("altgr cancel: lctl state->pending-released");
-                                lctl_state = LctlState::PendingReleased;
-                            }
-                            LctlState::Pending => {
-                                log::debug!("altgr cancel: lctl state->send");
-                                try_send_panic(
-                                    &process_tx,
-                                    KeyEvent::new(OsCode::KEY_LEFTCTRL, KeyValue::Press),
-                                );
-                                lctl_state = LctlState::None;
-                            }
-                            LctlState::PendingReleased => {
-                                log::debug!("altgr cancel: lctl state->send+release");
-                                try_send_panic(
-                                    &process_tx,
-                                    KeyEvent::new(OsCode::KEY_LEFTCTRL, KeyValue::Press),
-                                );
-                                try_send_panic(
-                                    &process_tx,
-                                    KeyEvent::new(OsCode::KEY_LEFTCTRL, KeyValue::Release),
-                                );
-                                lctl_state = LctlState::None;
-                            }
-                            _ => {}
-                        }
-                    }
-                    std::thread::sleep(time::Duration::from_millis(1));
-                }
-                Err(TryRecvError::Disconnected) => {
-                    panic!("channel disconnected")
-                }
-            }
-        }
-    });
+    set_win_altgr_behaviour(_cfg)?;
+    Ok(())
 }
 
 #[cfg(feature = "cmd")]
@@ -689,30 +445,6 @@ fn run_cmd(cmd_and_args: &'static [String]) -> std::thread::JoinHandle<()> {
             Err(e) => log::error!("Failed to execute cmd: {}", e),
         };
     })
-}
-
-fn set_altgr_behaviour(_cfg: &cfg::Cfg) -> Result<()> {
-    #[cfg(target_os = "windows")]
-    {
-        *ALTGR_BEHAVIOUR.lock() = {
-            const CANCEL: &str = "cancel-lctl-press";
-            const ADD: &str = "add-lctl-release";
-            match _cfg.items.get("windows-altgr") {
-                None => AltGrBehaviour::DoNothing,
-                Some(cfg_val) => match cfg_val.as_str() {
-                    CANCEL => AltGrBehaviour::CancelLctlPress,
-                    ADD => AltGrBehaviour::AddLctlRelease,
-                    _ => bail!(
-                        "Invalid value for windows-altgr: {}. Valid values are {},{}",
-                        cfg_val,
-                        CANCEL,
-                        ADD
-                    ),
-                },
-            }
-        };
-    }
-    Ok(())
 }
 
 #[cfg(feature = "cmd")]
