+++ src/kanata/windows.rs
@@ -1,0 +1,240 @@
+use anyhow::{bail, Result};
+use crossbeam_channel::{Receiver, Sender, TryRecvError};
+use parking_lot::Mutex;
+use std::collections::HashSet;
+use std::convert::TryFrom;
+use std::sync::Arc;
+use std::time;
+
+use crate::cfg;
+
+use super::*;
+
+#[derive(Debug, Clone, Copy, PartialEq)]
+pub enum AltGrBehaviour {
+    DoNothing,
+    CancelLctlPress,
+    AddLctlRelease,
+}
+
+static PRESSED_KEYS: Lazy<Mutex<HashSet<OsCode>>> = Lazy::new(|| Mutex::new(HashSet::new()));
+
+pub static ALTGR_BEHAVIOUR: Lazy<Mutex<AltGrBehaviour>> =
+    Lazy::new(|| Mutex::new(AltGrBehaviour::DoNothing));
+
+impl Kanata {
+    /// Initialize the callback that is passed to the Windows low level hook to receive key events
+    /// and run the native_windows_gui event loop.
+    pub fn event_loop(kanata: Arc<Mutex<Self>>, tx: Sender<KeyEvent>) -> Result<()> {
+        // Display debug and panic output when launched from a terminal.
+        unsafe {
+            use winapi::um::wincon::*;
+            if AttachConsole(ATTACH_PARENT_PROCESS) != 0 {
+                panic!("Could not attach to console");
+            }
+        };
+        native_windows_gui::init()?;
+        {
+            let mut mapped_keys = MAPPED_KEYS.lock();
+            *mapped_keys = kanata.lock().mapped_keys;
+        }
+
+        let (preprocess_tx, preprocess_rx) = crossbeam_channel::bounded(10);
+        start_event_preprocessor(preprocess_rx, tx);
+
+        // This callback should return `false` if the input event is **not** handled by the
+        // callback and `true` if the input event **is** handled by the callback. Returning false
+        // informs the callback caller that the input event should be handed back to the OS for
+        // normal processing.
+        let _kbhook = KeyboardHook::set_input_cb(move |input_event| {
+            if input_event.code as usize >= cfg::MAPPED_KEYS_LEN {
+                return false;
+            }
+            if !MAPPED_KEYS.lock()[input_event.code as usize] {
+                return false;
+            }
+
+            let mut key_event = match KeyEvent::try_from(input_event) {
+                Ok(ev) => ev,
+                _ => return false,
+            };
+
+            // Unlike Linux, Windows does not use a separate value for repeat. However, our code
+            // needs to differentiate between initial press and repeat press.
+            log::debug!("event loop: {:?}", key_event);
+            match key_event.value {
+                KeyValue::Release => {
+                    PRESSED_KEYS.lock().remove(&key_event.code);
+                }
+                KeyValue::Press => {
+                    let mut pressed_keys = PRESSED_KEYS.lock();
+                    if pressed_keys.contains(&key_event.code) {
+                        key_event.value = KeyValue::Repeat;
+                    } else {
+                        pressed_keys.insert(key_event.code);
+                    }
+                }
+                _ => {}
+            }
+
+            // Send input_events to the preprocessing loop. Panic if channel somehow gets full or if
+            // channel disconnects. Typing input should never trigger a panic based on the channel
+            // getting full, assuming regular operation of the program and some other bug isn't the
+            // problem. I've tried to crash the program by pressing as many keys on my keyboard at
+            // the same time as I could, but was unable to.
+            try_send_panic(&preprocess_tx, key_event);
+            true
+        });
+
+        // The event loop is also required for the low-level keyboard hook to work.
+        native_windows_gui::dispatch_thread_events();
+        Ok(())
+    }
+}
+
+fn try_send_panic(tx: &Sender<KeyEvent>, kev: KeyEvent) {
+    if let Err(e) = tx.try_send(kev) {
+        panic!("failed to send on channel: {:?}", e)
+    }
+}
+
+pub fn set_win_altgr_behaviour(cfg: &cfg::Cfg) -> Result<()> {
+    *ALTGR_BEHAVIOUR.lock() = {
+        const CANCEL: &str = "cancel-lctl-press";
+        const ADD: &str = "add-lctl-release";
+        match cfg.items.get("windows-altgr") {
+            None => AltGrBehaviour::DoNothing,
+            Some(cfg_val) => match cfg_val.as_str() {
+                CANCEL => AltGrBehaviour::CancelLctlPress,
+                ADD => AltGrBehaviour::AddLctlRelease,
+                _ => bail!(
+                    "Invalid value for windows-altgr: {}. Valid values are {},{}",
+                    cfg_val,
+                    CANCEL,
+                    ADD
+                ),
+            },
+        }
+    };
+    Ok(())
+}
+
+fn start_event_preprocessor(preprocess_rx: Receiver<KeyEvent>, process_tx: Sender<KeyEvent>) {
+    #[derive(Debug, Clone, Copy, PartialEq)]
+    enum LctlState {
+        Pressed,
+        Released,
+        Pending,
+        PendingReleased,
+        None,
+    }
+
+    std::thread::spawn(move || {
+        let mut lctl_state = LctlState::None;
+        loop {
+            match preprocess_rx.try_recv() {
+                Ok(kev) => match (*ALTGR_BEHAVIOUR.lock(), kev) {
+                    (AltGrBehaviour::DoNothing, _) => try_send_panic(&process_tx, kev),
+                    (
+                        AltGrBehaviour::AddLctlRelease,
+                        KeyEvent {
+                            value: KeyValue::Release,
+                            code: OsCode::KEY_RIGHTALT,
+                            ..
+                        },
+                    ) => {
+                        log::debug!("altgr add: adding lctl release");
+                        try_send_panic(&process_tx, kev);
+                        try_send_panic(
+                            &process_tx,
+                            KeyEvent::new(OsCode::KEY_LEFTCTRL, KeyValue::Release),
+                        );
+                        PRESSED_KEYS.lock().remove(&OsCode::KEY_LEFTCTRL);
+                    }
+                    (
+                        AltGrBehaviour::CancelLctlPress,
+                        KeyEvent {
+                            value: KeyValue::Press,
+                            code: OsCode::KEY_LEFTCTRL,
+                            ..
+                        },
+                    ) => {
+                        log::debug!("altgr cancel: lctl state->pressed");
+                        lctl_state = LctlState::Pressed;
+                    }
+                    (
+                        AltGrBehaviour::CancelLctlPress,
+                        KeyEvent {
+                            value: KeyValue::Release,
+                            code: OsCode::KEY_LEFTCTRL,
+                            ..
+                        },
+                    ) => match lctl_state {
+                        LctlState::Pressed => {
+                            log::debug!("altgr cancel: lctl state->released");
+                            lctl_state = LctlState::Released;
+                        }
+                        LctlState::Pending => {
+                            log::debug!("altgr cancel: lctl state->pending-released");
+                            lctl_state = LctlState::PendingReleased;
+                        }
+                        LctlState::None => try_send_panic(&process_tx, kev),
+                        _ => {}
+                    },
+                    (
+                        AltGrBehaviour::CancelLctlPress,
+                        KeyEvent {
+                            value: KeyValue::Press,
+                            code: OsCode::KEY_RIGHTALT,
+                            ..
+                        },
+                    ) => {
+                        log::debug!("altgr cancel: lctl state->none");
+                        lctl_state = LctlState::None;
+                        try_send_panic(&process_tx, kev);
+                    }
+                    (_, _) => try_send_panic(&process_tx, kev),
+                },
+                Err(TryRecvError::Empty) => {
+                    if *ALTGR_BEHAVIOUR.lock() == AltGrBehaviour::CancelLctlPress {
+                        match lctl_state {
+                            LctlState::Pressed => {
+                                log::debug!("altgr cancel: lctl state->pending");
+                                lctl_state = LctlState::Pending;
+                            }
+                            LctlState::Released => {
+                                log::debug!("altgr cancel: lctl state->pending-released");
+                                lctl_state = LctlState::PendingReleased;
+                            }
+                            LctlState::Pending => {
+                                log::debug!("altgr cancel: lctl state->send");
+                                try_send_panic(
+                                    &process_tx,
+                                    KeyEvent::new(OsCode::KEY_LEFTCTRL, KeyValue::Press),
+                                );
+                                lctl_state = LctlState::None;
+                            }
+                            LctlState::PendingReleased => {
+                                log::debug!("altgr cancel: lctl state->send+release");
+                                try_send_panic(
+                                    &process_tx,
+                                    KeyEvent::new(OsCode::KEY_LEFTCTRL, KeyValue::Press),
+                                );
+                                try_send_panic(
+                                    &process_tx,
+                                    KeyEvent::new(OsCode::KEY_LEFTCTRL, KeyValue::Release),
+                                );
+                                lctl_state = LctlState::None;
+                            }
+                            _ => {}
+                        }
+                    }
+                    std::thread::sleep(time::Duration::from_millis(1));
+                }
+                Err(TryRecvError::Disconnected) => {
+                    panic!("channel disconnected")
+                }
+            }
+        }
+    });
+}
